pyproject.toml
[tool.poetry]
name = "shanyraq"
version = "0.1.0"
description = ""
authors = ["Shyryn <shyryn1806@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.30"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"
bcrypt = "3.2.0"
passlib = "1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
